Refactor build scripts and replace Koin with Kodein DI
In an effort to improve the structure and efficiency of the code base, this commit focused primarily on adjusting the build setup and swapping the dependency injection framework. Koin has been replaced with Kodein DI throughout the code for superior performance and flexibility. Changes are also made to the build scripts (build.gradle.kts and settings.gradle.kts), removing redundant plugins and adding new ones as necessary. Additionally, the ViewModel retrieval mechanism has been updated in line with the transition to Kodein DI. A new file "kmp-resource-fix.gradle.kts" is added to fix resource generation issue in a multiplatform project.